# Draft: [POC][IMP] core: use RECURSIVE sub query for child_of