**Description of the issue/feature this PR addresses:**

It is a POC to improve the search with child_of. Today child_of is not made directly by SQL, it is prepare by ORM (https://github.com/odoo/odoo/blob/14.0/odoo/osv/expression.py#L526), but it can generate some horrible query.

This PR use the WITH RECURCIVE of postgres to do the same things.

Note : This PR include https://github.com/odoo/odoo/pull/89962 (to make the test)
Note 2: This PR is a quick POC

@rco-odoo @odony 

Exemple of ORM query (with 0,5 M of res.partner)
```python
self.env['account.move'].search(['&', ['move_type', '=', 'in_invoice'], '|', '|', '|', '|', ['name', 'ilike', 'a'], ['invoice_origin', 'ilike', 'a'], ['ref', 'ilike', 'a'], ['payment_reference', 'ilike', 'a'], ['partner_id', 'child_of', 'a']])
```

**Before :**
4 sql query with `['in', [lot of values]]`
```SQL
SELECT "res_partner".id FROM "res_partner" 
WHERE ("res_partner"."parent_id" in (29244,680426,672283,29281,29060,29162, ... 271696,47935,47937,47936,608349,608350,608351)) ORDER BY  "res_partner"."id"   
```

```SQL
SELECT "res_partner".id FROM "res_partner" 
WHERE ("res_partner"."parent_id" in (18,127,129,130,131,25024,26420,26421, ... ,684351,684353,684354)) ORDER BY  "res_partner"."id"   
```

```SQL
SELECT "res_partner".id FROM "res_partner" 
WHERE ("res_partner"."parent_id" in (27211,27212,30500,30501,30504,36658,75883, ....85689,115514,117246,119301,119336,129680,130251,204842,334772,481456,501000,609558,625311)) ORDER BY  "res_partner"."id"   
```

```SQL
SELECT "account_move".id FROM "account_move" 
WHERE (("account_move"."move_type" = 'in_invoice') AND (((((unaccent("account_move"."name"::text) ilike unaccent('%a%')) OR (unaccent("account_move"."invoice_origin"::text) ilike unaccent('%a%'))) OR (unaccent("account_move"."ref"::text) ilike unaccent('%a%'))) OR (unaccent("account_move"."payment_reference"::text) ilike unaccent('%a%'))) 
OR ("account_move"."partner_id" in (SELECT "res_partner".id FROM "res_partner" WHERE (("res_partner"."active" = true) 
AND ("res_partner"."id" in (1,7,10,11,16,17,18,103,126,127,129,130, ... 684352,684353,684354,684355))) AND ((("res_partner"."partner_share" IS NULL or "res_partner"."partner_share" = false ) OR ("res_partner"."company_id" in (1))) OR "res_partner"."company_id" IS NULL ))))) AND ("account_move"."company_id" IS NULL  OR ("account_move"."company_id" in (1))) 
ORDER BY  "account_move"."date" DESC,"account_move"."name" DESC,"account_move"."id" DESC  LIMIT 80 
```

Time to execute : 3.68 s
Time to execute search_count: 13.06 s

**After :**
1 sql query

```SQL
SELECT "account_move".id FROM "account_move" WHERE (("account_move"."move_type" = 'in_invoice') AND (((((unaccent("account_move"."name"::text) ilike unaccent('%a%')) OR (unaccent("account_move"."invoice_origin"::text) ilike unaccent('%a%'))) OR (unaccent("account_move"."ref"::text) ilike unaccent('%a%'))) OR (unaccent("account_move"."payment_reference"::text) ilike unaccent('%a%'))) OR ("account_move"."partner_id" in (
            WITH RECURSIVE subordinates AS (
            SELECT "res_partner".id FROM "res_partner" WHERE (("res_partner"."active" = true) AND ((((unaccent("res_partner"."display_name"::text) ilike unaccent('%a%')) OR (unaccent("res_partner"."ref"::text) ilike unaccent('%a%'))) OR (unaccent("res_partner"."email"::text) ilike unaccent('%a%'))) OR (unaccent("res_partner"."vat"::text) ilike unaccent('%a%')))) AND ((("res_partner"."partner_share" IS NULL or "res_partner"."partner_share" = false ) OR ("res_partner"."company_id" in (1))) OR "res_partner"."company_id" IS NULL )
            UNION
            SELECT alias_sub.id FROM "res_partner" alias_sub
            INNER JOIN subordinates s ON s.id = alias_sub.parent_id
            ) SELECT * FROM subordinates)))) AND ("account_move"."company_id" IS NULL  OR ("account_move"."company_id" in (1))) ORDER BY  "account_move"."date" DESC,"account_move"."name" DESC,"account_move"."id" DESC  LIMIT 80 

```

Time to execute : 0,96 s
Time to execute search_count : 10,8 s (Idea : the total of record is show in the tree view, it could be nice to compte in async to prevent a long freeze, because on large database the query with LIMI 80 is more faster than the query to count).


What do you think ?

--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr