@fmdl 2022-09-06
ping @rco-odoo