@rco-odoo 2022-11-18
This is an improvement, not a bugfix.
It therefore does not qualify for a stable branch https://github.com/odoo/odoo/wiki/Contributing#what-does-stable-mean