@fmdl 2022-11-18
@rco-odoo I know it is an improvement, I can move in master if it will be merged.